refactor(interfaces): chatbot as an interface of the system
