+++ src/interfaces/chatbot/index.test.ts
@@ -1,11 +1,11 @@
-import { chat } from './bot'
+import { chat } from '.'
 import { Chat } from 'twitch-js'
 
 describe('bot singleton', () => {
     describe('chat', () => {
         it('should aways be the same instance of chat element', async () => {
-            const oneChatInstance = await import('./bot').then(x => x.chat)
-            const otherChatInstance = await import('./bot').then(x => x.chat)
+            const oneChatInstance = await import('.').then(x => x.chat)
+            const otherChatInstance = await import('.').then(x => x.chat)
 
             expect(chat).toBe(oneChatInstance)
             expect(chat).toBe(otherChatInstance)
