+++ src/run-bot.ts
@@ -1,6 +1,6 @@
 import { Chat } from 'twitch-js'
 import onReceivePrivateMessages from './on-receive-messages'
-import { channel, chat } from './bot'
+import { channel, chat } from './interfaces/chatbot'
 
 const runBot = async (): Promise<void> => {
     chat.on(Chat.Events.PRIVATE_MESSAGE, onReceivePrivateMessages)
